+++ packages/realm_generator/lib/src/tmp/realm.schema.g.dart
@@ -1,0 +1,1 @@
+../../../../../generator/lib/src/tmp/realm.schema.g.dart