+++ packages/realm_generator/lib/src/tmp/test.generated.dart
@@ -1,0 +1,1 @@
+../../../../../generator/lib/src/tmp/test.generated.dart