+++ packages/realm_generator/test/realmobject.dart
@@ -1,0 +1,1 @@
+../../../generator/test/realmobject.dart