+++ packages/realm_generator/test/test.dart
@@ -1,0 +1,1 @@
+../../../generator/test/test.dart