+++ generator/CHANGELOG.md
@@ -1,7 +1,7 @@
 0.1.0-preview Release notes (yyyy-MM-dd)
 =============================================================
 ### Enhancements
-* The initial preview version of the Realm Generator for Flutter and Dart.
+* The initial preview version of the Realm Generator for Realm lutter and Realm Dart SDKs
 
 ### Fixed
 * -